+++ tests/test_sdk.py
@@ -3,6 +3,7 @@ import pytest
 import uuid
 import pandas as pd
 
+from decimal import Decimal
 from tests import API_KEY
 from ntropy_sdk import (
     SDK,
@@ -362,7 +363,7 @@ def test_batch(sdk):
     resp, status = batch.poll()
     assert status == "finished" and resp[0].merchant == "Amazon Web Services"
 
-    batch = Batch(sdk, batch.batch_id)
+    batch = Batch(sdk=sdk, batch_id=batch.batch_id)
     resp, status = batch.poll()
     assert status == "finished" and resp[0].merchant == "Amazon Web Services"
 
@@ -393,7 +394,7 @@ def test_train_custom_model(sdk):
     _, status, _ = model.poll()
     assert status in ["enriching", "training", "queued"] and model.is_synced()
 
-    m = Model(sdk, model_name, poll_interval=1)
+    m = Model(sdk=sdk, model_name=model_name, poll_interval=1)
     _, status, _ = model.poll()
     assert status in ["enriching", "training", "queued"] and m.is_synced()
 
@@ -451,7 +452,7 @@ def test_train_custom_model_df(sdk):
     _, status, _ = model.poll()
     assert status in ["enriching", "training", "queued"] and model.is_synced()
 
-    m = Model(sdk, model_name, poll_interval=1)
+    m = Model(sdk=sdk, model_name=model_name, poll_interval=1)
     _, status, _ = model.poll()
     assert status in ["enriching", "training", "queued"] and m.is_synced()
 
@@ -476,3 +477,18 @@ def test_train_custom_model_df(sdk):
     )[0]
 
     assert "supermarket" in e.labels
+
+
+def test_numerical_support():
+    tx = Transaction(
+        amount=Decimal(24.56),
+        description="TARGET T- 5800 20th St 11/30/19 17:32",
+        entry_type="debit",
+        date="2012-12-10",
+        account_holder_id="1",
+        iso_currency_code="USD",
+        mcc="5432",
+    )
+
+    assert isinstance(tx.amount, float)
+    assert isinstance(tx.mcc, int)